add line break to menu options
A line break has been added to menu options to prevent confusion with selecting the right value for menu options. Depending on the width of the console pane, text can be wrapped onto different lines. In some cases, this ends up with license number being in the same column as the menu choice number. This fixes that.